Feat: Added views to frontend. By selecting header button, view will change.
