+++ src/app/page.tsx
@@ -1,65 +1,25 @@
 "use client";
 
 import Header from "@/components/custom/header";
+import ImageFeed from "@/components/custom/ImageFeed";
+import ImageGenerate from "@/components/custom/ImageGenerate";
+import UserProfile from "@/components/custom/UserProfile";
 import { useState } from "react";
 
 export default function Home() {
-  const [inputText, setInputText] = useState("");
-  const [isLoading, setIsLoading] = useState(false);
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true);
-
-    try {
-      const response = await fetch("/api/generate-image", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ text: inputText }),
-      });
-
-      const data = await response.json();
-      console.log(data);
-      setInputText("");
-    } catch (error) {
-      console.error("Error:", error);
-    } finally {
-      setIsLoading(false);
-    }
-  };
+  const [view, setView] = useState<'feed' | 'generate' | 'profile'>("feed");
 
   return (
     // TODO: Update the UI here to show the images generated
     
-    <div className="min-h-screen flex flex-col justify-between p-8">
-      <Header />
-      <main className="flex-1">
+    <div className="flex flex-col min-h-screen p-8">
+      <Header setView={setView} />
+      <main className="container mx-auto px-4 py-8 flex-grow flex flex-col">
         {/* Main content can go here */}
+        {view === 'feed' && <ImageFeed />}
+        {view === 'generate' && <ImageGenerate />}
+        {view == 'profile' && <UserProfile />}
       </main>
-
-      <footer className="w-full max-w-3xl mx-auto">
-        <form onSubmit={handleSubmit} className="w-full">
-          <div className="flex gap-2">
-            <input
-              type="text"
-              value={inputText}
-              onChange={e => setInputText(e.target.value)}
-              className="flex-1 p-3 rounded-lg bg-black/[.05] dark:bg-white/[.06] border border-black/[.08] dark:border-white/[.145] focus:outline-none focus:ring-2 focus:ring-black dark:focus:ring-white"
-              placeholder="Describe the image you want to generate..."
-              disabled={isLoading}
-            />
-            <button
-              type="submit"
-              disabled={isLoading}
-              className="px-6 py-3 rounded-lg bg-foreground text-background hover:bg-[#383838] dark:hover:bg-[#ccc] transition-colors disabled:opacity-50"
-            >
-              {isLoading ? "Generating..." : "Generate"}
-            </button>
-          </div>
-        </form>
-      </footer>
     </div>
   );
 }
