+++ src/components/custom/ImageFeed.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+
+export default function ImageFeed() {
+
+    return (
+        <div className="w-full">
+            <h1>Take a look at other pictures!</h1>
+        </div>
+    );
+}
