+++ src/components/custom/UserProfile.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+
+
+export default function UserProfile() {
+
+    return (<div className="w-full">
+        <h1>Look at your Info!</h1>
+    </div>);
+}
